add integration tests for build script updates, move integration tests into their own files
