+++ build_runner/test/generate/build_test.dart
@@ -3,13 +3,10 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:build/build.dart';
 import 'package:glob/glob.dart';
-import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
-import 'package:test_descriptor/test_descriptor.dart' as d;
 
 import 'package:build_runner/build_runner.dart';
 import 'package:build_runner/src/asset_graph/exceptions.dart';
@@ -581,33 +578,6 @@ void main() {
       expect(writer.assets.containsKey(aCloneNode.id), isFalse);
     });
 
-    test('invalidates graph if build script updates', () async {
-      var graph = await AssetGraph.build([], new Set(), 'a', null);
-      var aId = makeAssetId('a|web/a.txt');
-      var aCopyNode = new GeneratedAssetNode(
-          0, aId, false, true, makeAssetId('a|web/a.txt.copy'),
-          lastKnownDigest: computeDigest('a'));
-      graph.add(aCopyNode);
-      var aNode =
-          makeAssetNode('a|web/a.txt', [aCopyNode.id], computeDigest('a'));
-      graph.add(aNode);
-
-      var writer = new InMemoryRunnerAssetWriter();
-
-      /// Spoof the `package:test/test.dart` import and pretend its newer than
-      /// the current graph to cause a rebuild.
-      await writer.writeAsString(makeAssetId('test|lib/test.dart'), '');
-      await testActions([
-        copyABuildAction
-      ], {
-        'a|web/a.txt': 'a',
-        'a|web/a.txt.copy': 'a',
-        'a|$assetGraphPath': JSON.encode(graph.serialize()),
-      }, outputs: {
-        'a|web/a.txt.copy': 'a',
-      }, writer: writer);
-    }, skip: 'Need to move this to an integration test');
-
     test('no outputs if no changed sources', () async {
       var graph = await AssetGraph.build([], new Set(), 'a', null);
       var aId = makeAssetId('a|web/a.txt');
@@ -663,221 +633,6 @@ void main() {
           writer: writer,
           writeToCache: true,
           packageGraph: packageGraph);
-    });
-  });
-
-  group('build integration tests', () {
-    group('findAssets', () {
-      setUp(() async {
-        await d.dir('a', [
-          await pubspec('a', currentIsolateDependencies: [
-            'build',
-            'build_runner',
-            'build_test',
-            'glob'
-          ]),
-          d.dir('tool', [
-            d.file('build.dart', '''
-import 'package:build_runner/build_runner.dart';
-import 'package:build_test/build_test.dart';
-import 'package:glob/glob.dart';
-
-main() async {
-  await build(
-    [new BuildAction(new GlobbingBuilder(new Glob('**.txt')), 'a')]);
-}
-''')
-          ]),
-          d.dir('web', [
-            d.file('a.globPlaceholder'),
-            d.file('a.txt', ''),
-            d.file('b.txt', ''),
-          ]),
-        ]).create();
-
-        await pubGet('a');
-
-        // Run a build and validate the output.
-        var result = await runDart('a', 'tool/build.dart');
-        expect(result.exitCode, 0, reason: result.stderr as String);
-        await d.dir('a', [
-          d.dir('web', [d.file('a.matchingFiles', 'a|web/a.txt\na|web/b.txt')])
-        ]).validate();
-      });
-
-      test('picks up new files that match the glob', () async {
-        // Add a new file matching the glob.
-        await d.dir('a', [
-          d.dir('web', [d.file('c.txt', '')])
-        ]).create();
-
-        // Run a new build and validate.
-        var result = await runDart('a', 'tool/build.dart');
-        expect(result.exitCode, 0, reason: result.stderr as String);
-        expect(result.stdout, contains('with 1 outputs'));
-        await d.dir('a', [
-          d.dir('web', [
-            d.file('a.matchingFiles', 'a|web/a.txt\na|web/b.txt\na|web/c.txt')
-          ])
-        ]).validate();
-      });
-
-      test('picks up deleted files that match the glob', () async {
-        // Delete a file matching the glob.
-        var aTxtFile = new File(p.join(d.sandbox, 'a', 'web', 'a.txt'));
-        aTxtFile.deleteSync();
-
-        // Run a new build and validate.
-        var result = await runDart('a', 'tool/build.dart');
-        expect(result.exitCode, 0, reason: result.stderr as String);
-        expect(result.stdout, contains('with 1 outputs'));
-        await d.dir('a', [
-          d.dir('web', [d.file('a.matchingFiles', 'a|web/b.txt')])
-        ]).validate();
-      });
-
-      test(
-          'doesn\'t cause new builds for files that don\'t match '
-          'any globs', () async {
-        // Add a new file not matching the glob.
-        await d.dir('a', [
-          d.dir('web', [d.file('c.other', '')])
-        ]).create();
-
-        // Run a new build and validate.
-        var result = await runDart('a', 'tool/build.dart');
-        expect(result.exitCode, 0, reason: result.stderr as String);
-        expect(result.stdout, contains('with 0 outputs'));
-        await d.dir('a', [
-          d.dir('web', [d.file('a.matchingFiles', 'a|web/a.txt\na|web/b.txt')])
-        ]).validate();
-      });
-
-      test('doesn\'t cause new builds for file changes', () async {
-        // Change a file matching the glob.
-        await d.dir('a', [
-          d.dir('web', [d.file('a.txt', 'changed!')])
-        ]).create();
-
-        // Run a new build and validate.
-        var result = await runDart('a', 'tool/build.dart');
-        expect(result.exitCode, 0, reason: result.stderr as String);
-        expect(result.stdout, contains('with 0 outputs'));
-        await d.dir('a', [
-          d.dir('web', [d.file('a.matchingFiles', 'a|web/a.txt\na|web/b.txt')])
-        ]).validate();
-      });
-    });
-
-    group('findAssets with no initial output', () {
-      setUp(() async {
-        await d.dir('a', [
-          await pubspec('a', currentIsolateDependencies: [
-            'build',
-            'build_runner',
-            'build_test',
-            'glob'
-          ]),
-          d.dir('tool', [
-            d.file('build.dart', '''
-import 'package:build_runner/build_runner.dart';
-import 'package:build_test/build_test.dart';
-import 'package:glob/glob.dart';
-
-main() async {
-  await build(
-    [new BuildAction(new OverDeclaringGlobbingBuilder(
-        new Glob('**.txt')), 'a')]);
-}
-
-class OverDeclaringGlobbingBuilder extends GlobbingBuilder {
-  OverDeclaringGlobbingBuilder(Glob glob) : super(glob);
-
-  @override
-  Future build(BuildStep buildStep) async {
-    var assets = await buildStep.findAssets(glob).toList();
-    // Only output if we have a 'web/b.txt' file.
-    if (assets.any((id) => id.path == 'web/b.txt')) {
-      await super.build(buildStep);
-    }
-  }
-}
-''')
-          ]),
-          d.dir('web', [
-            d.file('a.globPlaceholder'),
-            d.file('a.txt', ''),
-          ]),
-        ]).create();
-
-        await pubGet('a');
-
-        // Run a build and validate the output.
-        var result = await runDart('a', 'tool/build.dart');
-        expect(result.exitCode, 0, reason: result.stderr as String);
-        expect(result.stdout, contains('with 0 outputs'));
-        await d.dir('a', [
-          d.dir('web', [d.nothing('a.matchingFiles')])
-        ]).validate();
-      });
-
-      test('picks up new files that match the glob', () async {
-        // Add a new file matching the glob which causes a real output.
-        await d.dir('a', [
-          d.dir('web', [d.file('b.txt', '')])
-        ]).create();
-
-        // Run a new build and validate.
-        var result = await runDart('a', 'tool/build.dart');
-        expect(result.exitCode, 0, reason: result.stderr as String);
-        expect(result.stdout, contains('with 1 outputs'));
-        await d.dir('a', [
-          d.dir('web', [d.file('a.matchingFiles', 'a|web/a.txt\na|web/b.txt')])
-        ]).validate();
-      });
-    });
-  });
-
-  group('regression tests', () {
-    test(
-        'checking for existing outputs works with deleted '
-        'intermediate outputs', () async {
-      await d.dir('a', [
-        await pubspec('a', currentIsolateDependencies: [
-          'build',
-          'build_runner',
-          'build_test',
-          'glob'
-        ]),
-        d.dir('tool', [
-          d.file('build.dart', '''
-import 'package:build_runner/build_runner.dart';
-import 'package:build_test/build_test.dart';
-
-main() async {
-  await build([
-    new BuildAction(new CopyBuilder(), 'a'),
-    new BuildAction(new CopyBuilder(), 'a', inputs: ['**.txt.copy']),
-  ]);
-}
-''')
-        ]),
-        d.dir('web', [
-          d.file('a.txt', 'a'),
-          d.file('a.txt.copy.copy', 'a'),
-        ]),
-      ]).create();
-
-      await pubGet('a');
-
-      var result = await runDart('a', 'tool/build.dart');
-
-      expect(result.exitCode, isNot(0),
-          reason: 'build should fail due to conflicting outputs');
-      expect(
-          result.stderr,
-          allOf(contains('Found 1 outputs already on disk'),
-              contains('a|web/a.txt.copy.copy')));
     });
   });
 }
